feat(storybook-addon): rename from plugin to addon
