--- packages/storybook-plugin/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/storybook-plugin';
